fix(sql-store): make the bigint backfill terminate and search staged indexes (#549)
* fix(vue): open the pinned page range on loadMore

Vue's paginated core re-keyed the open-ended subscription to the pinned
key instead of closing it, so the bounded `(null, cursor]` range was never
subscribed: page one kept an open `LIMIT n` query, duplicated or dropped
rows across the boundary on writes, and could never split. Svelte, Solid
and Angular already did this right.

Across all four adapters:
- paginated APIs accept `onError` and expose `error`; a page failure no
  longer freezes the feed at `LoadingMore`
- `hydratePreloaded` forwards `onError` (Vue, Svelte, Solid)
- the value resets when reactive args change instead of showing the
  previous args' rows until the first frame
- Svelte keeps the server error `code`; agent stores accept `onError`

replica: diff incoming frames by primary key instead of re-inserting every
row per frame into an unbounded event log; bound the log (1000, drop
oldest); page the `/since` catch-up and drop the unused `/range`.

BREAKING CHANGE: `EventLogDOClient.getSince` returns a page; `getRange`
is removed; paginated result types gain `error`.

(cherry picked from commit 03606cf126b7020bbdf4414abc14a5f47e83047d)

* fix(replica): bound the catch-up loops and key rows the wire actually sends

`GET /since` validated `seq` and `limit` as merely finite. `?limit=1.5` reaches SQLite's
`LIMIT ?`, which rejects it, so a malformed request answers 500 instead of the 400 it has
earned; `?seq=1.5` moves the `seq >= ?` boundary past a real entry while reporting the
page the caller asked for. Both are now checked with `Number.isSafeInteger`, keeping the
existing range bounds.

The mirror snapshot diff classified a `bigint` primary key as un-keyed. That is what the
wire decoder hands back for an int64 column, so an id-typed key lands there: the row was
inserted on every frame (never recorded as known, so never "unchanged") and could never be
deleted (never known, so never missing). Key it like the other two.

Both catch-up loops ran until the log was exhausted, which never happens against a log
written faster than it is read: `EventsSync`'s poll cycle never returns, so `#inFlight` —
which every concurrent `sync()` awaits — never settles, and `MaterializerRuntime.initialize`
never finishes startup. Each now spends a bounded page budget per call and leaves the
remainder to the next one, at the advanced watermark, so progress is unchanged.

`EventsSync` also replayed events through `applyEvents` twice whenever a batch's replay
succeeded and the run failed downstream, in `getTableDiffs` or partway through the mirror
fan-out. The watermark correctly holds there, so the retry re-fetches the same batch — but
only `getTableDiffs` is required to be idempotent; `applyEvents` mutates a state machine
with nothing to roll back. Track the highest sequence already replayed and hand the state
machine only what is past it. A replay that threw is not recorded, so that batch is still
retried whole.

(cherry picked from commit b737038898eff776010209b3de30da64d39da0f4)

* fix(sql-store): make the bigint backfill terminate and search staged indexes

- a staged search index on an empty `.global()` table never got a state
  row, so every search returned `SEARCH_INDEX_BUILDING` until an operator
  ran the backfill by hand; skip staged indexes only when the table has rows
- the bigint key rewrite re-scanned every row on every isolate cold start
  (`LENGTH()` cannot use an index); a per-table completion marker in
  `__lunora_migration_state` makes a converted table one lookup
- the rewrite UPDATE now carries a compare-and-swap so a user write that
  commits mid-conversion is not reverted to the legacy value
- one bigint codec: the decoder lives beside the encoder in shard-engine
- `undeclaredField` uses `Object.hasOwn` (a `constructor` key passed)
- a durable shape joining a `.memory()` relation is refused like a shape
  over one
- migrations split out of `ctx-db.ts`; `normalizeCapacity` has one home

BREAKING CHANGE: the recorded search-index profile now carries the field
as well as the analyzer, so every deployed search index rebuilds once on
the first migration after this ships, on both planes. The rebuild is in
place and coverage stays latched, so it costs one backfill walk per index,
not an outage. `BIGINT_KEY_NEGATIVE`/`BIGINT_KEY_NON_NEGATIVE` are no
longer exported; `decodeBigintSqlKey` is.

(cherry picked from commit 303a6c3e79b41100a085ce1da6a038db8cef8a9d)

---------
