+++ packages/observability/__tests__/span-buffer.test.ts
@@ -67,6 +67,35 @@ describe("spanBuffer", () => {
     });
 });
 
+describe("spanBuffer capacity normalization", () => {
+    it("falls back to the default for a capacity that would truncate to zero", () => {
+        expect.assertions(2);
+
+        // `> 0` accepted this and `Math.trunc` then made it 0, so the ring
+        // evicted every span it was handed — trace capture silently off.
+        const buffer = new SpanBuffer(0.5);
+
+        buffer.push(span());
+
+        expect(buffer.size).toBe(1);
+        expect(buffer.dropped).toBe(0);
+    });
+
+    it("falls back to the default for a non-finite capacity", () => {
+        expect.assertions(1);
+
+        // Infinity truncates to itself, removing the memory bound the ring
+        // exists to impose on a buffer that lives as long as the DO does.
+        const buffer = new SpanBuffer(Number.POSITIVE_INFINITY);
+
+        for (let index = 0; index < 1200; index += 1) {
+            buffer.push(span({ spanId: String(index) }));
+        }
+
+        expect(buffer.size).toBeLessThan(1200);
+    });
+});
+
 describe("foldTraces", () => {
     it("returns nothing for an empty buffer", () => {
         expect.assertions(2);
