+++ packages/sql-store/src/ctx-db.ts
@@ -30,7 +30,6 @@ import type {
     AggregateResult,
     AggregateTally,
     CdcChange,
-    ColumnMetaLike,
     CrossShardReadArgs,
     DatabaseWriterLike,
     GroupByEntry,
@@ -48,7 +47,6 @@ import type {
     TriggerEventLike,
     TriggerOpLike,
     TriggerTimingLike,
-    ValidatorLike,
     WhereInput,
     WhereSqlStrategy,
 } from "@lunora/shard-engine";
@@ -90,7 +88,6 @@ import {
     rankTableName,
     readAggregateValue,
     relationHooks,
-    renderSql,
     resolveRankPartition,
     resolveRankSeekTuple,
     resolveRelationPredicates,
@@ -110,14 +107,26 @@ import { sql } from "drizzle-orm";
 
 import { evictOldestEntry } from "../../../shared/evict-oldest";
 import { decodeWire, encodeWire, needsWireEncoding } from "../../../shared/wire-codec";
+import { runSqlAggregateMigrations, runSqlGlobalTableMigrations, runSqlRankMigrations } from "./ctx-db-migrations";
 import type { SearchStage } from "./ctx-db-search";
 import { createSearchSync, runSqlSearch, runSqlSearchMigrations } from "./ctx-db-search";
-import { migrateSearchState, readSearchBackfillState, writeSearchBackfillState } from "./ctx-db-search-state";
+import { migrateSearchState } from "./ctx-db-search-state";
 import type { SqlDialect } from "./dialect";
 import { createPointReadBatcher } from "./point-read-batcher";
 import type { SqlCtxExec } from "./sql-exec";
-import { columnRefSql, createIndexIfNotExists, decodeRow, decodeRows, forEachRowPaged, queryAll, queryBatch, queryRun, serializeColumnValue } from "./sql-exec";
-import { BIGINT_KEY_LENGTH, bigintSqlKey, effectiveColumnKind } from "./value-codec";
+import {
+    columnRefSql,
+    createIndexIfNotExists,
+    decodeRow,
+    decodeRows,
+    forEachRowPaged,
+    queryAll,
+    queryBatch,
+    queryRun,
+    serializeColumnValue,
+    tableColumns,
+} from "./sql-exec";
+import { bigintSqlKey, effectiveColumnKind } from "./value-codec";
 
 /** Order fields that already provide a stable tiebreak (no extra `id` term needed). */
 const ID_ORDER_FIELDS = new Set(["_id", "id"]);
@@ -294,21 +303,6 @@ interface SqlCtxDbOptions {
 
 /** Cap on re-entrant trigger writes before we treat it as a self-triggering loop. */
 const MAX_TRIGGER_DEPTH = 50;
-
-/** A table's fields paired with their column meta, skipping fields that declare none. */
-const tableColumns = (definition: TableDefinitionLike): [string, ColumnMetaLike][] => {
-    const columns: [string, ColumnMetaLike][] = [];
-
-    for (const [field, validator] of Object.entries(definition.shape)) {
-        const column = validator._meta?.column;
-
-        if (column) {
-            columns.push([field, column]);
-        }
-    }
-
-    return columns;
-};
 
 /**
  * Fill any field absent from `document` that declares a `.default()` literal or
@@ -699,217 +693,6 @@ const encodeRankCursor = (cursorValues: ReadonlyArray<unknown>): string => {
 };
 
 /**
- * SQLite affinity for a column. Resolves the *effective* validator kind (so
- * `v.optional(inner)` stores as `inner` would) and defers to the shared dialect
- * (`@lunora/d1/dialect`) — the same mapping the `lunora migrate generate` SQL
- * emitter uses, so auto-provisioned and hand-migrated tables stay identical.
- */
-const globalColumnAffinity = (validator: ValidatorLike, dialect: SqlDialect): string => dialect.columnType(effectiveColumnKind(validator));
-
-/** Build the column DDL for a global table as a drizzle `SQL`: framework columns plus a typed column per declared field. */
-const globalTableColumnsDdl = (tableName: string, definition: SchemaLike["tables"][string], dialect: SqlDialect): SQL => {
-    const fieldColumns: SQL[] = [];
-
-    for (const [field, validator] of Object.entries(definition.shape)) {
-        if (!validator._meta?.column) {
-            continue;
-        }
-
-        // Required, non-optional fields get NOT NULL; optional ones stay nullable
-        // so an insert that omits them can't trip a constraint.
-        const notNull = validator._meta.column.notNull && validator.kind !== "optional" ? " NOT NULL" : "";
-
-        fieldColumns.push(sql`${sql.identifier(field)} ${sql.raw(`${globalColumnAffinity(validator, dialect)}${notNull}`)}`);
-    }
-
-    const frameworkColumns = dialect.frameworkColumns().map((column) => sql`${sql.identifier(column.name)} ${sql.raw(column.type)}`);
-    const total = frameworkColumns.length + fieldColumns.length;
-
-    // `VALIDATION_ERROR`, not `INTERNAL`: a table too wide is the schema
-    // author's input, and an internal-coded error has its message replaced with
-    // "Internal error" on the way out — redacting the one sentence that says
-    // what to do. `ensureMigrated` does not cache the rejection, so every
-    // request re-runs this; an opaque 500 forever is a bad way to learn a table
-    // has too many columns.
-    if (dialect.maxTableColumns !== undefined && total > dialect.maxTableColumns) {
-        throw new LunoraError(
-            "VALIDATION_ERROR",
-            `@lunora/sql-store: global table "${tableName}" needs ${String(total)} columns, over this engine's ${String(dialect.maxTableColumns)}-column limit — split the table, or move the extra fields into one object field`,
-        );
-    }
-
-    return sql.join([...frameworkColumns, ...fieldColumns], sql`, `);
-};
-
-/**
- * The default total order every read here ends with, appended to a declared
- * index so the index can answer the sort instead of the engine sorting every
- * match into a temp B-tree.
- *
- * Same defect, same fix, as `INDEX_SORT_KEYS` in `@lunora/shard-engine`'s
- * `ctx-db-migrations.ts` — this backend simply never got it. `compileOrderBySql`
- * below used to say so in its own doc. Measured on `node:sqlite`, 50k rows,
- * `WHERE status = ? ORDER BY _creationTime, id LIMIT 21`: 134.0us against a
- * fields-only index, 11.3us with the sort keys on it, and the `USE TEMP B-TREE
- * FOR ORDER BY` step disappears from the plan.
- *
- * NOT appended to a UNIQUE index: `(email, _creationTime, id)` is unique for
- * every row, so the constraint would silently stop rejecting duplicates — data
- * corruption rather than a slow query. Same split as the DO twin.
- */
-const indexSortKeys = (dialect: SqlDialect): SQL | undefined => {
-    // Not on an engine that needs a key prefix to index text. MySQL declares
-    // `id VARCHAR(768)` — 768 utf8mb4 characters is 3072 bytes, which is exactly
-    // InnoDB's whole-index key limit on its own — so appending it to ANY other
-    // column raises ER_TOO_LONG_KEY and the migration fails outright. Prefixing
-    // it (`id(191)`) would create, but MySQL cannot satisfy an ORDER BY from a
-    // prefixed column, so the index would cost writes and buy nothing. The
-    // engine that actually backs `.global()` is D1 (SQLite); Postgres indexes
-    // text directly and gets the fix too.
-    if (dialect.indexKeyPrefix?.("string") !== undefined) {
-        return undefined;
-    }
-
-    return sql`${columnRefSql("_creationTime")}, ${columnRefSql("id")}`;
-};
-
-/**
- * Drop `name` when the engine already holds an index by that name built from a
- * DIFFERENT column list.
- *
- * `CREATE INDEX IF NOT EXISTS` does not replace a differing definition and does
- * not complain, so a database provisioned before {@link indexSortKeys} existed
- * would keep its filter-only index forever and never see the improvement. The
- * DO twin (`dropIndexIfShapeChanged`) exists for exactly this.
- *
- * SQLite only — that is the engine `.global()` tables run on (D1), and it is the
- * one whose catalog echoes the CREATE statement back verbatim, so the comparison
- * is a string compare rather than a per-engine column-list reconstruction. On
- * Postgres/MySQL a pre-existing index of the old shape is left alone: the new
- * shape reaches fresh databases, and an existing one keeps working, slower.
- */
-const dropIndexIfShapeChanged = async (
-    exec: SqlCtxExec,
-    dialect: SqlDialect,
-    spec: { columns: SQL; name: string; table: string; unique: boolean },
-): Promise<void> => {
-    if (dialect.name !== "sqlite") {
-        return;
-    }
-
-    const held = await queryAll(exec, dialect, sql`SELECT sql FROM sqlite_master WHERE type = 'index' AND name = ${spec.name} AND tbl_name = ${spec.table}`);
-    const current = held[0]?.["sql"];
-
-    // No row, or an implicit index the engine created for a constraint (`sql` is
-    // NULL there) — nothing of ours to replace.
-    if (typeof current !== "string") {
-        return;
-    }
-
-    // Compare the parenthesised column list rather than the whole statement, so
-    // this stays insensitive to how the surrounding DDL is spelled.
-    const columnsOf = (statement: string): string | undefined => {
-        const open = statement.indexOf("(");
-
-        return open === -1 ? undefined : statement.slice(open, statement.lastIndexOf(")") + 1);
-    };
-
-    const unique = spec.unique ? sql`UNIQUE ` : sql``;
-    const wanted = columnsOf(
-        renderSql(dialect.name, sql`CREATE ${unique}INDEX ${sql.identifier(spec.name)} ON ${sql.identifier(spec.table)} (${spec.columns})`).sql,
-    );
-
-    if (wanted === undefined || wanted === columnsOf(current)) {
-        return;
-    }
-
-    // A UNIQUE index is dropped only once we know the new shape can actually be
-    // created. Dropping first and letting the follow-up `CREATE UNIQUE INDEX`
-    // fail on rows that are duplicates under the NEW column set leaves the table
-    // with no constraint at all — and the failed migration re-runs and re-fails
-    // on every wake, so the gap does not close on its own. Refusing here keeps
-    // the old constraint in force and names what has to be de-duplicated.
-    //
-    // There is a TOCTOU window between this probe and the create, which is
-    // acceptable: this runs at provisioning time, only when an index's declared
-    // fields actually changed, and losing the race costs a failed migration
-    // rather than a silently unprotected table.
-    if (spec.unique) {
-        const duplicates = await queryAll(
-            exec,
-            dialect,
-            sql`SELECT ${spec.columns} FROM ${sql.identifier(spec.table)} GROUP BY ${spec.columns} HAVING COUNT(*) > 1 LIMIT 1`,
-        );
-
-        if (duplicates.length > 0) {
-            throw new LunoraError(
-                "INTERNAL",
-                `unique index "${spec.name}" on "${spec.table}" cannot be re-created with its new column list: existing rows are duplicates under it. De-duplicate the table with a data migration first; the previous index is left in place.`,
-            );
-        }
-    }
-
-    await queryRun(exec, dialect, sql`DROP INDEX IF EXISTS ${sql.identifier(spec.name)}`);
-};
-
-/** Create a global table's declared secondary indexes and its synthesized `.unique()` column indexes. */
-const createGlobalTableIndexes = async (exec: SqlCtxExec, tableName: string, definition: SchemaLike["tables"][string], dialect: SqlDialect): Promise<void> => {
-    // Index column reference as drizzle SQL, with a key prefix where the engine
-    // demands it (MySQL can't index its now-unbounded TEXT string columns without
-    // one). Framework columns (id/_creationTime — absent from `shape`) are already
-    // index-safe types, so they get no prefix.
-    const indexRef = (field: string): SQL => {
-        const reference = columnRefSql(field);
-        const validator = definition.shape[field];
-        const prefix = validator && dialect.indexKeyPrefix ? dialect.indexKeyPrefix(effectiveColumnKind(validator)) : undefined;
-
-        return prefix === undefined ? reference : sql`${reference}(${sql.raw(String(prefix))})`;
-    };
-
-    const sortKeys = indexSortKeys(dialect);
-
-    for (const index of definition.indexes) {
-        const fields = sql.join(
-            index.fields.map((field) => indexRef(field)),
-            sql`, `,
-        );
-        const unique = index.unique ?? false;
-        const spec = {
-            columns: unique || sortKeys === undefined ? fields : sql`${fields}, ${sortKeys}`,
-            name: `${tableName}_${index.name}`,
-            table: tableName,
-            unique,
-        };
-
-        // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the shared D1 connection.
-        await dropIndexIfShapeChanged(exec, dialect, spec);
-        // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the shared D1 connection.
-        await createIndexIfNotExists(exec, dialect, spec);
-    }
-
-    // The DEFAULT total order, for the reads that name no index at all — a bare
-    // `findMany({ limit })` or `paginate()`. The table above declares only `id`
-    // as its primary key, so nothing indexed `ORDER BY _creationTime, id` and the
-    // engine read the whole table into a temp B-tree to return the first page.
-    // The DO twin creates the same index on every row table.
-    if (sortKeys !== undefined) {
-        await createIndexIfNotExists(exec, dialect, { columns: sortKeys, name: `${tableName}__by_creation`, table: tableName, unique: false });
-    }
-
-    // `.unique()` columns synthesize a UNIQUE index so the engine enforces the
-    // constraint (the write layer maps breaches to ConflictError), mirroring the
-    // DO twin's `migrateSecondaryIndexes`.
-    for (const [field, column] of tableColumns(definition)) {
-        if (!column.unique) {
-            continue;
-        }
-
-        // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the shared D1 connection.
-        await createIndexIfNotExists(exec, dialect, { columns: indexRef(field), name: `${tableName}_unique_${field}`, table: tableName, unique: true });
-    }
-};
-
-/**
  * Refuse a SQL-side reduce or group over a column stored as an order-preserving
  * key rather than as its value.
  *
@@ -919,15 +702,11 @@ const createGlobalTableIndexes = async (exec: SqlCtxExec, tableName: string, def
  * characters of padding. All three look like answers, and `SUM` past 2^53 used
  * instead to escape as a raw driver `RangeError`.
  *
- * The maintained `__agg_` companion answers these numerically, so the error names
- * it rather than just refusing. Exact per CONTRIBUTION, not per total: `__value__`
- * is a DOUBLE, and `applyAggregateDelta` accumulates coerced values into it — so
- * a `sum` of individually safe `bigint` amounts loses integer precision once the
- * running total passes 2^53, and a `min`/`max` over values past it compares the
- * nearest doubles. A caller who needs an exact total over that range reads the
- * rows and reduces them in JS, where the values are `bigint` again.
- *
- * Applied at every SQL-reducing entry point —
+ * The maintained `__agg_` companion is what the error names instead. It is exact
+ * per contribution — `coerceAggregateNumber` refuses any single `bigint` past
+ * 2^53 outright — but its running total accumulates in a REAL column, so a sum
+ * of in-range values can still cross 2^53 and round there. The message says so
+ * rather than promising exactness the companion cannot give. Applied at every SQL-reducing entry point —
  * `aggregate`'s scan and both halves of `groupBy` — matching the shard twin's
  * `assertReducibleBySql`, which shipped guarding one and not its sibling.
  * @throws LunoraError `BAD_REQUEST` when `field` is stored as an order-preserving key
@@ -938,7 +717,7 @@ const assertReducibleBySql = (definition: SchemaLike["tables"][string], field: s
     if (validator !== undefined && effectiveColumnKind(validator) === "bigint") {
         throw new LunoraError(
             "BAD_REQUEST",
-            `${label}: "${field}" is stored as an order-preserving key, which SQL cannot reduce or group — declare an aggregateIndex covering this (by, field, op) so the maintained companion answers it`,
+            `${label}: "${field}" is stored as an order-preserving key, which SQL cannot reduce or group — declare an aggregateIndex covering this (by, field, op) so the maintained companion answers it instead (its running total is a REAL, so it stays exact only while the total is inside 2^53)`,
         );
     }
 };
@@ -956,436 +735,6 @@ const assertGroupByReducibleBySql = (
 
     if (agg.field !== undefined) {
         assertReducibleBySql(definition, agg.field, `groupBy(${tableName}, { agg: { op: "${agg.op}", field: "${agg.field}" } })`);
-    }
-};
-
-/**
- * Rows converted per pass by {@link rewriteLegacyBigintColumns}; keeps one
- * statement's bound-parameter count inside D1's budget.
- *
- * Rendered into the `LIMIT` inline (`sql.raw`) rather than bound: mysql2's
- * prepared-statement path rejects a placeholder there with
- * `Incorrect arguments to mysqld_stmt_execute`, which took the whole MySQL
- * provisioning pass down. It is a module constant, so nothing caller-supplied
- * reaches the statement text.
- */
-const BIGINT_REWRITE_PAGE = 100;
-
-/**
- * Migration-state key for one table's `bigint` re-encoding pass.
- *
- * Recorded in the same reserved table the search backfill uses — it is a
- * `key → {cursor, done, profile}` map, and this pass needs exactly that. The
- * `:` namespace cannot collide with a search companion name (`companionFor`
- * emits `<table>__search_<index>`).
- */
-const bigintRewriteStateKey = (tableName: string): string => `bigint-rewrite:${tableName}`;
-
-/**
- * Stamped alongside a recorded completion, so widening {@link BIGINT_KEY_LENGTH}
- * one day re-runs the pass instead of reading a marker for the old width.
- */
-const BIGINT_REWRITE_PROFILE = `bigint-key/${String(BIGINT_KEY_LENGTH)}`;
-
-/**
- * The declared fields of `table` that are stored as a `bigint` key.
- */
-const bigintFields = (definition: SchemaLike["tables"][string]): string[] =>
-    Object.entries(definition.shape)
-        .filter(([, validator]) => validator._meta?.column !== undefined && effectiveColumnKind(validator) === "bigint")
-        .map(([field]) => field);
-
-/**
- * The `SET` assignments that convert one row's legacy `bigint` columns, empty
- * when every one of them already holds a key.
- *
- * A value this encoding cannot hold — non-numeric text no encoder here ever
- * wrote, or a magnitude past the 39-digit ceiling — is left exactly as stored
- * rather than aborting the pass; the keyset cursor still steps over it.
- */
-const legacyBigintAssignments = (row: Record<string, unknown>, fields: ReadonlyArray<string>): SQL[] => {
-    const assignments: SQL[] = [];
-
-    for (const field of fields) {
-        const raw = row[field];
-
-        // "Already a key", not "already 40 characters" — the same distinction the
-        // `legacy` predicate makes. A key is a `"0"`/`"1"` sign character plus 39
-        // digits; `"-"` plus a 39-digit magnitude is 40 characters too, and a
-        // length-only test skipped exactly those, leaving a convertible negative
-        // stored as decimal text that `eq` no longer matches.
-        if (typeof raw !== "string" || (raw.length === BIGINT_KEY_LENGTH && !raw.startsWith("-"))) {
-            continue;
-        }
-
-        try {
-            assignments.push(sql`${columnRefSql(field)} = ${bigintSqlKey(BigInt(raw))}`);
-        } catch {
-            // Not a value this encoding can hold. Left as stored.
-        }
-    }
-
-    return assignments;
-};
-
-/**
- * Rows still holding decimal text, expressed so the rows the encoding can never
- * hold stop matching permanently.
- *
- * A convertible legacy value is at most a signed 39-digit decimal — 40
- * characters with the sign — so anything longer is past the encoding's digit
- * ceiling and is left as stored. That leaves the one overlap with a real key, which is
- * also 40 characters: a key's first character is its `"0"`/`"1"` sign, never
- * `"-"`, so `SUBSTR(col, 1, 1) = '-'` separates them. A plain `LENGTH(col) <> 40`
- * matched neither, which silently skipped every negative 39-digit value.
- */
-const legacyBigintPredicate = (fields: ReadonlyArray<string>): SQL =>
-    sql.join(
-        fields.map(
-            (field) =>
-                sql`(${columnRefSql(field)} IS NOT NULL AND (LENGTH(${columnRefSql(field)}) < ${BIGINT_KEY_LENGTH} OR (LENGTH(${columnRefSql(field)}) = ${BIGINT_KEY_LENGTH} AND SUBSTR(${columnRefSql(field)}, 1, 1) = ${"-"})))`,
-        ),
-        sql` OR `,
-    );
-
-/** One `UPDATE` per row of a page that has something to convert; rows without a usable `id`, and values the encoding cannot hold, are skipped. */
-const legacyBigintUpdates = (rows: ReadonlyArray<Record<string, unknown>>, fields: ReadonlyArray<string>, tableName: string): SQL[] => {
-    const updates: SQL[] = [];
-
-    for (const row of rows) {
-        const { id } = row;
-        const assignments = typeof id === "string" ? legacyBigintAssignments(row, fields) : [];
-
-        if (assignments.length > 0) {
-            updates.push(sql`UPDATE ${sql.identifier(tableName)} SET ${sql.join(assignments, sql`, `)} WHERE ${columnRefSql("id")} = ${id}`);
-        }
-    }
-
-    return updates;
-};
-
-/**
- * Rewrite any `v.bigint()` column still holding the plain decimal text an
- * earlier build stored into the order-preserving key {@link bigintSqlKey} now
- * writes.
- *
- * This is a **storage-format migration, and it is not optional**. Decimal text
- * is exact for `=` but sorts `"9"` after `"10"`, so every range filter,
- * `ORDER BY`, page cursor and `MIN`/`MAX` over such a column returned the wrong
- * rows. The fix changes what is written — which means a table holding both forms
- * has a worse problem than the one it started with: `where: { n: { eq: 10n } }`
- * binds the key and no longer matches a row stored as `"10"`. Converting the
- * stragglers is what keeps that from being a silent read break, so it runs from
- * the same provisioning pass that creates the table.
- *
- * Self-terminating, and recorded once done. It pages on a keyset cursor over
- * `id`, so a value the encoding cannot hold is stepped over rather than retried
- * forever within a pass — but the cursor is not durable, so without a marker the
- * next `ensureMigrated` restarted at the top. `ensureMigrated` is memoised per
- * ctx-db (per REQUEST on a Hyperdrive binding), and the probe is a scan: on a
- * converted table it matches nothing only after reading every row. So completion
- * goes in the shared migration-state table and the steady state is one
- * primary-key lookup, not a full-table scan per request.
- *
- * The predicate skips what the encoding provably cannot hold — anything longer
- * than a signed 39-digit decimal — so an oversized legacy value stops matching
- * permanently rather than pinning the pass open. It does still match a sign
- * character: `"-"` plus 39 digits is 40 characters, so a plain `LENGTH <> 40`
- * left exactly the negative 39-digit values unconverted, and a converted key
- * never starts with `-`.
- */
-const rewriteLegacyBigintColumns = async (
-    exec: SqlCtxExec,
-    tableName: string,
-    definition: SchemaLike["tables"][string],
-    dialect: SqlDialect,
-): Promise<void> => {
-    const fields = bigintFields(definition);
-
-    if (fields.length === 0) {
-        return;
-    }
-
-    const stateKey = bigintRewriteStateKey(tableName);
-    const recorded = await readSearchBackfillState(exec, dialect, stateKey);
-
-    if (recorded.done && recorded.profile === BIGINT_REWRITE_PROFILE) {
-        return;
-    }
-
-    const legacy = legacyBigintPredicate(fields);
-    const selected = sql.join([columnRefSql("id"), ...fields.map((field) => columnRefSql(field))], sql`, `);
-    let cursor = "";
-
-    for (;;) {
-        // eslint-disable-next-line no-await-in-loop -- keyset pages run sequentially on the single shared connection; each page's cursor comes from the previous one.
-        const rows = await queryAll(
-            exec,
-            dialect,
-            sql`SELECT ${selected} FROM ${sql.identifier(tableName)} WHERE (${legacy}) AND ${columnRefSql("id")} > ${cursor} ORDER BY ${columnRefSql("id")} LIMIT ${sql.raw(String(BIGINT_REWRITE_PAGE))}`,
-        );
-
-        if (rows.length === 0) {
-            break;
-        }
-
-        // eslint-disable-next-line no-await-in-loop -- one round trip per page; the rows are keyed by distinct `id`, so order across them doesn't matter.
-        await queryBatch(exec, dialect, legacyBigintUpdates(rows, fields, tableName));
-
-        const last = rows.at(-1)?.["id"];
-
-        if (typeof last !== "string") {
-            // The cursor cannot advance, so this pass did NOT reach the end of
-            // the table and must not be recorded as converged.
-            return;
-        }
-
-        cursor = last;
-    }
-
-    await writeSearchBackfillState(exec, dialect, stateKey, undefined, true, BIGINT_REWRITE_PROFILE);
-};
-
-/**
- * Add the columns a `.global()` table is missing.
- *
- * `CREATE TABLE IF NOT EXISTS` alone leaves an existing table exactly as it is,
- * so adding a field to a shipped schema provisioned nothing: every `insert` died
- * with the driver's own `table p has no column named slug` — untyped, and never
- * mentioning a remedy — while reads and unrelated patches kept working, so the
- * deploy looked half-healthy. Both sibling migrations in this package already
- * reshape their own tables (`runSqlAggregateMigrations` PRAGMA-checks and
- * `ALTER`s, `migrateSearchState` `ALTER`s under a try/catch); only the
- * user-facing table was left out.
- *
- * The probe is one `SELECT <every declared column> … LIMIT 0` — dialect-blind,
- * no catalog member on {@link SqlDialect}, and on the overwhelmingly common
- * no-drift path it is a single statement that reads no rows. Only when that
- * fails does it cost one probe per column to find which are missing.
- *
- * Added columns are always **nullable**, whatever the field declares: existing
- * rows have no value for a field that did not exist, and `ADD COLUMN … NOT NULL`
- * without a default is rejected outright on a non-empty table by all three
- * engines. The declared `notNull` is still enforced on write by the validator.
- * Additive only, like the `CREATE` it follows — a retype or a drop still needs an
- * explicit migration.
- */
-const alterGlobalTableDrift = async (exec: SqlCtxExec, tableName: string, definition: SchemaLike["tables"][string], dialect: SqlDialect): Promise<void> => {
-    const fields = Object.entries(definition.shape).filter(([, validator]) => validator._meta?.column !== undefined);
-
-    if (fields.length === 0) {
-        return;
-    }
-
-    const probe = async (columns: ReadonlyArray<string>): Promise<boolean> => {
-        try {
-            await queryAll(
-                exec,
-                dialect,
-                sql`SELECT ${sql.join(
-                    columns.map((column) => columnRefSql(column)),
-                    sql`, `,
-                )} FROM ${sql.identifier(tableName)} LIMIT 0`,
-            );
-
-            return true;
-        } catch {
-            return false;
-        }
-    };
-
-    if (await probe(fields.map(([field]) => field))) {
-        return;
-    }
-
-    for (const [field, validator] of fields) {
-        // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the single shared connection; each probe gates its own ALTER.
-        if (await probe([field])) {
-            continue;
-        }
-
-        // eslint-disable-next-line no-await-in-loop -- same connection, and a failure here must surface rather than race the next ALTER.
-        await queryRun(
-            exec,
-            dialect,
-            sql`ALTER TABLE ${sql.identifier(tableName)} ADD COLUMN ${sql.identifier(field)} ${sql.raw(globalColumnAffinity(validator, dialect))}`,
-        );
-    }
-};
-
-/**
- * Auto-provision every `.global()` table from the schema: `CREATE TABLE IF NOT
- * EXISTS` with the physical `id`/`_creationTime` columns plus a typed column per
- * declared field, then its secondary and `.unique()` indexes. This is the D1
- * twin of `@lunora/do`'s `runShardMigrations` (which self-creates shard-local
- * tables) — it makes the schema the single source of truth for global tables
- * too, so a fresh database serves them without a hand-applied migration. The
- * column set and dialect match exactly what this module reads and writes
- * (`columnRef`, `serializeColumnValue`, `decodeGlobalRow`).
- *
- * Idempotent (`CREATE TABLE/INDEX IF NOT EXISTS`); additive only — it never
- * drops or retypes an existing column, so destructive schema changes still need
- * an explicit migration.
- */
-const runSqlGlobalTableMigrations = async (exec: SqlCtxExec, schema: SchemaLike, dialect: SqlDialect): Promise<void> => {
-    // The bigint re-encoding pass below records its completion there, and this is
-    // also the entry point a dev host calls on its own — "the migration throws
-    // unless you happened to run the search one first" is not a contract.
-    // Idempotent, so the later call from `ensureMigrated` costs nothing.
-    await migrateSearchState(exec, dialect);
-
-    for (const [tableName, definition] of Object.entries(schema.tables)) {
-        if (definition.shardMode?.kind !== "global") {
-            continue;
-        }
-
-        const columns = globalTableColumnsDdl(tableName, definition, dialect);
-
-        // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the single shared D1 connection; the table must exist before its indexes below.
-        await queryRun(exec, dialect, sql`CREATE TABLE IF NOT EXISTS ${sql.identifier(tableName)} (${columns})`);
-        // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially; the table must carry every declared column before its indexes reference them.
-        await alterGlobalTableDrift(exec, tableName, definition, dialect);
-        // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially; indexes follow the table.
-        await createGlobalTableIndexes(exec, tableName, definition, dialect);
-        // eslint-disable-next-line no-await-in-loop -- runs on the same connection, after the columns exist.
-        await rewriteLegacyBigintColumns(exec, tableName, definition, dialect);
-    }
-};
-
-/**
- * Materialize the `__agg_<index>` companion tables for every declared
- * `aggregateIndex` on a global table. Global tables in Lunora ship their own
- * DDL — counter tables are opt-in so production hosts can decide where they
- * live. Tests and dev hosts can call this once after their schema migration to
- * unlock O(1) counts.
- *
- * Idempotent (`CREATE TABLE IF NOT EXISTS`).
- */
-const runSqlAggregateMigrations = async (exec: SqlCtxExec, schema: SchemaLike, dialect: SqlDialect): Promise<void> => {
-    const { integer, key, real } = dialect.companionTypes;
-
-    for (const [tableName, definition] of Object.entries(schema.tables)) {
-        const indexes = definition.aggregateIndexes;
-
-        if (!indexes || indexes.length === 0) {
-            continue;
-        }
-
-        for (const index of indexes) {
-            const aggTable = aggregateTableName(tableName, index.name);
-
-            // `__value__` is op-aware now (count / running sum / extreme — NULL
-            // for an empty min/max group) and `__count__` tracks the row count
-            // (avg divisor + empty-group detection). It is nullable; the pre-
-            // reducer-aware shape declared it `NOT NULL`.
-            // eslint-disable-next-line no-await-in-loop -- DDL statements run sequentially on the single shared connection.
-            await queryRun(
-                exec,
-                dialect,
-                sql`CREATE TABLE IF NOT EXISTS ${sql.identifier(aggTable)} (${sql.identifier("__key__")} ${sql.raw(key)} PRIMARY KEY, ${sql.identifier("__value__")} ${sql.raw(real)}, ${sql.identifier("__count__")} ${sql.raw(integer)} NOT NULL DEFAULT 0)`,
-            );
-
-            // Alpha-era companion-rebuild caveat (SQLite/D1 only): a binding that
-            // materialized this table before `__count__` existed gets the column
-            // added here (defaulted 0). `CREATE TABLE IF NOT EXISTS` won't reshape
-            // an existing table, so we pragma-check then ALTER. Fresh PG/MySQL
-            // tables are created with `__count__` already, so this legacy reshape
-            // is skipped off SQLite.
-            if (dialect.name === "sqlite") {
-                // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the shared connection.
-                const columns = await queryAll(exec, dialect, sql`PRAGMA table_info(${sql.identifier(aggTable)})`);
-
-                if (!columns.some((column) => column["name"] === "__count__")) {
-                    // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the shared connection.
-                    await queryRun(
-                        exec,
-                        dialect,
-                        sql`ALTER TABLE ${sql.identifier(aggTable)} ADD COLUMN ${sql.identifier("__count__")} ${sql.raw(integer)} NOT NULL DEFAULT 0`,
-                    );
-                }
-            }
-        }
-    }
-};
-
-/**
- * A rank btree index column. MySQL can't index a full VARCHAR(768)/TEXT column in
- * a *composite* index (3072-byte key limit), so VARCHAR/TEXT key columns get a
- * 191-char utf8mb4 prefix (keeps several columns under the cap); SQLite/Postgres
- * — and number/real columns everywhere — index in full.
- */
-const rankIndexColumn = (dialect: SqlDialect, column: string, direction: "ASC" | "DESC", needsPrefix: boolean): SQL => {
-    const reference = dialect.name === "mysql" && needsPrefix ? sql`${sql.identifier(column)}(191)` : sql`${sql.identifier(column)}`;
-
-    return sql`${reference} ${sql.raw(direction)}`;
-};
-
-/** The rank btree key tuple in sort order: `__partition__`, the sort columns, then `__id__` — each prefixed where MySQL's type demands it. */
-const rankBtreeColumns = (dialect: SqlDialect, index: RankIndexDefinitionLike, definition: SchemaLike["tables"][string]): SQL[] => {
-    // __partition__/__id__ are the VARCHAR(768) `key` type → always prefixed on MySQL.
-    const columns: SQL[] = [rankIndexColumn(dialect, "__partition__", "ASC", true)];
-
-    for (const [i, sortKey] of index.sortBy.entries()) {
-        const validator = definition.shape[sortKey.field];
-        const needsPrefix = validator !== undefined && dialect.indexKeyPrefix?.(effectiveColumnKind(validator)) !== undefined;
-
-        columns.push(rankIndexColumn(dialect, sortColumnName(i), sortKey.direction === "desc" ? "DESC" : "ASC", needsPrefix));
-    }
-
-    columns.push(rankIndexColumn(dialect, "__id__", "ASC", true));
-
-    return columns;
-};
-
-/** Each rank sort column is typed by its source field's kind (the same type + serialized form the main table uses), so it accepts the stored sort key and orders correctly. A generic BLOB would reject the value on Postgres (BYTEA is strict). */
-const rankSortColumnDefs = (dialect: SqlDialect, index: RankIndexDefinitionLike, definition: SchemaLike["tables"][string]): SQL[] =>
-    index.sortBy.map((sortKey, i) => {
-        const validator = definition.shape[sortKey.field];
-        const columnType = dialect.columnType(validator ? effectiveColumnKind(validator) : undefined);
-
-        return sql`${sql.identifier(sortColumnName(i))} ${sql.raw(columnType)}`;
-    });
-
-/**
- * Materialize the `__rank_<index>` companion tables for every declared
- * `rankIndex` on a global table. Mirrors `runSqlAggregateMigrations` — same
- * opt-in pattern so production hosts decide whether to spend the DDL.
- *
- * Idempotent (`CREATE TABLE IF NOT EXISTS` + `createIndexIfNotExists`).
- */
-const runSqlRankMigrations = async (exec: SqlCtxExec, schema: SchemaLike, dialect: SqlDialect): Promise<void> => {
-    const { key } = dialect.companionTypes;
-
-    for (const [tableName, definition] of Object.entries(schema.tables)) {
-        const indexes = definition.rankIndexes;
-
-        if (!indexes || indexes.length === 0) {
-            continue;
-        }
-
-        for (const index of indexes) {
-            const rankTable = rankTableName(tableName, index.name);
-            const sortColumnDefs = rankSortColumnDefs(dialect, index, definition);
-            const columnPart = sortColumnDefs.length > 0 ? sql`, ${sql.join(sortColumnDefs, sql`, `)}` : sql``;
-
-            // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the shared connection; the table must exist before its index below.
-            await queryRun(
-                exec,
-                dialect,
-                sql`CREATE TABLE IF NOT EXISTS ${sql.identifier(rankTable)} (${sql.identifier("__id__")} ${sql.raw(key)} PRIMARY KEY, ${sql.identifier("__partition__")} ${sql.raw(key)} NOT NULL${columnPart})`,
-            );
-
-            const orderedColumns = rankBtreeColumns(dialect, index, definition);
-            const btreeName = `${tableName}__rank_${index.name}__btree`;
-
-            // eslint-disable-next-line no-await-in-loop -- DDL runs sequentially on the shared D1 connection (the CREATE INDEX follows its CREATE TABLE).
-            await createIndexIfNotExists(exec, dialect, {
-                columns: sql.join(orderedColumns, sql`, `),
-                name: btreeName,
-                table: rankTable,
-                unique: false,
-            });
-        }
     }
 };
 
@@ -4143,17 +3492,7 @@ const createSqlCtxDb = (options: SqlCtxDbOptions): DatabaseWriterLike => {
     return writer;
 };
 
-export {
-    createSqlCtxDb,
-    readSqlCdcChangedTables,
-    readSqlCdcChanges,
-    readSqlCdcFloor,
-    runSqlAggregateMigrations,
-    runSqlCdcMigration,
-    runSqlGlobalTableMigrations,
-    runSqlRankMigrations,
-    sweepSqlCdcRetention,
-};
+export { createSqlCtxDb, readSqlCdcChangedTables, readSqlCdcChanges, readSqlCdcFloor, runSqlCdcMigration, sweepSqlCdcRetention };
 export { backfillSqlSearchIndexes, runSqlSearchMigrations } from "./ctx-db-search";
 export type { SqlCtxDbOptions };
